# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A comprehensive API for managing cargo and vessel tracking in logistics operations",
        "title": "DEUS Logistics API",
        "contact": {
            "name": "API Support"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/vessels": {
            "get": {
                "description": "Retrieve all vessels in the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vessel"
                ],
                "summary": "List all vessels",
                "responses": {
                    "200": {
                        "description": "List of vessels",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/http.VesselResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new vessel with specified capacity and location",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vessel"
                ],
                "summary": "Create a new vessel",
                "parameters": [
                    {
                        "description": "Vessel creation payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.CreateVesselRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Vessel created successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/http.VesselResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request or validation failed",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/vessels/{id}": {
            "get": {
                "description": "Retrieve a specific vessel by its UUID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vessel"
                ],
                "summary": "Get vessel by ID",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Vessel ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vessel found",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/http.VesselResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid vessel ID format",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Vessel not found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/vessels/{id}/location": {
            "patch": {
                "description": "Update the current location of a vessel",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vessel"
                ],
                "summary": "Update vessel location",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Vessel ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Location update payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.UpdateVesselLocationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Location updated successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/http.VesselResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request or validation failed",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Vessel not found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "http.CreateVesselRequest": {
            "type": "object",
            "required": [
                "capacity",
                "current_location",
                "name"
            ],
            "properties": {
                "capacity": {
                    "type": "number",
                    "example": 500
                },
                "current_location": {
                    "type": "string",
                    "example": "Port of Singapore"
                },
                "name": {
                    "type": "string",
                    "example": "MV Ocean Navigator"
                }
            }
        },
...
        "http.UpdateVesselLocationRequest": {
            "type": "object",
            "required": [
                "current_location"
            ],
            "properties": {
                "current_location": {
                    "type": "string",
                    "example": "Port of Rotterdam"
                }
            }
        },
        "http.VesselResponse": {
            "type": "object",
            "properties": {
                "capacity": {
                    "type": "number",
                    "example": 500
                },
                "created_at": {
                    "type": "string",
                    "example": "2024-01-01T08:00:00Z"
                },
                "current_location": {
                    "type": "string",
                    "example": "Port of Singapore"
                },
                "id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174001"
                },
                "name": {
                    "type": "string",
                    "example": "MV Ocean Navigator"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2024-01-15T14:45:00Z"
                }
            }
        },
        "response.ErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/response.ErrorDetail"
                }
            }
        },
        "response.Meta": {
            "type": "object",
            "properties": {
                "request_id": {
                    "type": "string"
                }
            }
        },
        "response.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "meta": {
                    "$ref": "#/definitions/response.Meta"
                }
            }
        }
    }
}